Change the way hook.py creates _*Hook objects
Instead of having each decorator create *Hook object, it will just pass the type to _add_hook() (which it already did), and _add_hook will create a new hook if there isn't already one there.

This is in preperation for adding an add_hook function to PluginManager, but it also cleans up hook.py a bit too.